requirements.txt
aioredis==1.3.1
amqp==5.0.5
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
beautifulsoup4==4.9.3
billiard==3.6.3.0
cachetools==4.2.1
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
Django==3.1.6
django-addanother==2.0.2
django-phonenumber-field==5.0.0
django-polymorphic==3.0.0
django-rest-polymorphic==0.1.9
django-storages==1.11.1
Django-Verify-Email==0.0.5
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
docker==4.4.4
google-api-core==1.26.0
google-auth==1.27.0
google-cloud-core==1.6.0
google-cloud-storage==1.36.1
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
hiredis==2.0.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
kombu==5.0.2
lxml==4.6.3
mosspy==1.0.9
msgpack==1.0.2
packaging==20.9
phonenumbers==8.12.18
prompt-toolkit==3.0.16
protobuf==3.15.3
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
pyparsing==2.4.7
pytz==2021.1
redis==3.5.3
requests==2.25.1
rsa==4.7.2
service-identity==18.1.0
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
Twisted==21.2.0
txaio==21.2.1
urllib3==1.26.3
vine==5.0.0
wcwidth==0.2.5
websocket-client==0.57.0
zope.interface==5.4.0
